+++ components/Wysiwyg/Buttons/index.tsx
@@ -1,6 +1,8 @@
-import React, { useRef } from "react";
+import React, { ChangeEvent, useRef } from "react";
 
-export function ImageButton(props: { onChange: any }) {
+export function ImageButton(props: {
+  onChange: React.ChangeEventHandler<HTMLInputElement> | undefined;
+}) {
   const input = useRef<HTMLInputElement>(null);
   return (
     <button className="p-3" onClick={() => input.current?.click()}>
