+++ components/Wysiwyg/Editor.tsx
@@ -1,17 +1,17 @@
 "use client";
 
-import styles from "./styles/editor.module.css";
+import styles from "../../styles/editors/editor.module.css";
 import { useEditor, EditorContent } from "@tiptap/react";
 import StarterKit from "@tiptap/starter-kit";
 import Highlight from "@tiptap/extension-highlight";
 import Typography from "@tiptap/extension-typography";
 import Image from "@tiptap/extension-image";
 import CharacterCount from "@tiptap/extension-character-count";
-import { Emoji } from "./Extensions/Emoji";
-import MainMenu from "./Menus/MainMenu";
-import FloatMenu from "./Menus/FloatMenu";
-import PopMenu from "./Menus/PopMenu";
-import WordCount from "./Utils/WordCount";
+import { Emoji } from "./extensions/emoji";
+import MainMenu from "./menus/main-menu";
+import FloatMenu from "./menus/float-menu";
+import PopMenu from "./menus/pop-menu";
+import WordCount from "./utils/word-count";
 
 export default () => {
   const limit = 2000000;
@@ -30,7 +30,7 @@ export default () => {
   });
 
   return (
-    <div className="w-full relative aspect-square ">
+    <div className="w-full relative border border-solid border-slate-300 aspect-square ">
       <MainMenu editor={editor} />
       <FloatMenu editor={editor} />
       <PopMenu editor={editor} />
