+++ components/Wysiwyg/Utils/word-count.tsx
