+++ components/default-editor/preview.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from "react";
-import marked from "./Utils/marked";
-import preview from "./Styles/preview.module.css";
+import marked from "./utils/marked";
+import preview from "../../styles/editors/preview.module.css";
 
 interface Props {
   code: string;
@@ -14,7 +14,7 @@ export default function Preview({ code }: Props) {
     })();
   }, [code]);
   return (
-    <div className={"bg-slate-200 w-[200px] h-[200px] " + preview.wrapper}>
+    <div className={"bg-slate-200 p-3 w-full min-h-[200px] " + preview.wrapper}>
       <div dangerouslySetInnerHTML={{ __html: $code }}></div>
       {/* {$code} */}
     </div>
