+++ components/default-editor/utils/hljs-config.ts
